add banner to index page
