+++ src/components/@auth/@activate/ActivateScene.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { bool, string, func, shape } from 'prop-types'
 import { Typography } from '@material-ui/core'
 import AuthCard from 'src/components/@auth/Card/AuthCard'
-import Loading from 'components/Loading'
+import Loading from 'components/elements/Loading'
 import connector from './connector'
 
 class ActivateScene extends React.Component {
