--- src/components/Alert/connector.js
@@ -1,13 +1,0 @@
-import { bindActionCreators } from 'redux'
-import { connect } from 'react-redux'
-import actions from 'src/redux/ui/alert/action'
-
-const mapStateToProps = store => store.ui.alert
-
-const mapDispatchToProps = dispatch => ({
-  actions: {
-    alert: bindActionCreators(actions, dispatch),
-  },
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)
