--- src/components/Background/index.jsx
@@ -1,25 +1,0 @@
-import React from 'react'
-import { object, node } from 'prop-types'
-import connector from './connector'
-
-class Background extends React.Component {
-  componentDidUpdate() {
-    const { layout } = this.props
-    if (layout.background) {
-      document.body.style.backgroundImage = `url(${layout.background})`
-    } else {
-      document.body.style.backgroundImage = ''
-    }
-  }
-
-  render() {
-    return this.props.children
-  }
-}
-
-Background.propTypes = {
-  layout: object.isRequired,
-  children: node.isRequired,
-}
-
-export default connector(Background)
