--- src/components/common/DevTools.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const DevTools = () => process.env.NODE_ENV === 'development' ? <div id="devtools" /> : ''
-
-DevTools.propTypes = {}
-
-export default DevTools
