+++ src/components/elements/DevTools.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const DevTools = () => process.env.NODE_ENV === 'development' ? <div id="devtools" /> : ''
+
+DevTools.propTypes = {}
+
+export default DevTools
